Removed business logic from repository and tested the project

+++ Order.Application/Services/Products/ProductService.cs
@@ -1,7 +1,9 @@
 ﻿using AutoMapper;
 using Order.Application.Dtos.Product;
 using Order.Application.Interfaces;
+using Order.Application.Interfaces.Helper;
 using Order.Core.Entities;
+using Order.Core.Exceptions;
 using Order.Core.Interfaces;
 
 namespace Order.Application.Services.Products;
@@ -9,23 +11,23 @@ namespace Order.Application.Services.Products;
 public class ProductService : IProductService
 {
     private readonly IGenericRepository<Product> _repository;
+    private readonly IValidator<AddProductDto> _validator;
+    private readonly InventoryService _inventoryService;
     private readonly IMapper _mapper;
     
-    public ProductService(IGenericRepository<Product> repository, IMapper mapper)
+    public ProductService(IGenericRepository<Product> repository, IValidator<AddProductDto> validator, InventoryService inventoryService, IMapper mapper)
     {
         _repository = repository;
+        _validator = validator;
+        _inventoryService = inventoryService;
         _mapper = mapper;
     }
     
     public async Task Add(AddProductDto req)
     {
-        if (req.Price <= 0)
-            throw new ArgumentException("Price must be greater than 0");
-        if (req.Stock <= 0)
-            throw new ArgumentException("Stock must be greater than 0");
+        _validator.Validate(req); 
         
-        var products = await _repository.FilterAsync(p => p.Name == req.Name);
-        var productInDb = products.FirstOrDefault();
+        var productInDb = await _repository.GetFirstOrDefaultAsync(p => p.Name == req.Name);
         
         if (productInDb != null)
         {
@@ -38,27 +40,41 @@ public class ProductService : IProductService
         await _repository.AddAsync(product);
     }
 
-    public async Task<ProductDto> GetById(int id)
-    {
-        var product = await _repository.GetByIdAsync(id);
-        return _mapper.Map<ProductDto>(product);
-    }
-
     public async Task<List<ProductDto>> GetAll()
     {
         var products = await _repository.GetAllAsync();
         return _mapper.Map<List<ProductDto>>(products);
     }
 
-    public async Task DecreaseStock(int productId, int quantity)
+    public async Task<List<ProductDto>> SearchProductByName(string name)
     {
-        var product = await _repository.GetByIdAsync(productId);
-        product.Stock -= quantity;
+        var products = await _repository.FilterAsync(p => p.Name.Contains(name));
+        if (!products.Any())
+        {
+            var allProducts = await _repository.GetAllAsync();
+            return _mapper.Map<List<ProductDto>>(allProducts);
+        }
+        return _mapper.Map<List<ProductDto>>(products);
+    }
+    
+    public async Task<ProductDto> GetById(int id)
+    {
+        var product = await _repository.GetByIdAsync(id);
+        if (product == null)
+            throw new NotFoundException("Product not found");
+        return _mapper.Map<ProductDto>(product);
+    }
+
+    public async Task IncreaseStock(UpdateProductStockDto req)
+    {
+        await _inventoryService.IncreaseStock(req.productId, req.quantity);
     }
 
     public async Task Delete(int id)
     {
         var product = await _repository.GetByIdAsync(id);
+        if (product == null)
+            throw new NotFoundException("Product not found");
         await _repository.DeleteAsync(product);
     }
 }
