+++ Order.Application/Services/Users/UserService.cs
@@ -1,7 +1,9 @@
 ﻿using AutoMapper;
 using Order.Application.Dtos.User;
 using Order.Application.Interfaces;
+using Order.Application.Interfaces.Helper;
 using Order.Core.Entities;
+using Order.Core.Exceptions;
 using Order.Core.Interfaces;
 
 namespace Order.Application.Services.Users;
@@ -10,37 +12,34 @@ public class UserService : IUserService
 {
     private readonly IGenericRepository<User> _repository;
     private readonly IGenericRepository<Cart> _cartService;
+    private readonly IValidator<AddUserDto> _validator;
     private readonly IMapper _mapper;
     
-    public UserService(IGenericRepository<User> repository, IGenericRepository<Cart> cartService, IMapper mapper)
+    public UserService(IGenericRepository<User> repository, IGenericRepository<Cart> cartService, IValidator<AddUserDto> validator, IMapper mapper)
     {
         _repository = repository;
         _cartService = cartService;
+        _validator = validator;
         _mapper = mapper;
     }
     
-    public async Task Add(AddUserDto request)
+    public async Task Add(AddUserDto req)
     {
-        request.Email = request.Email.Trim().ToLower();
+        req.Email = req.Email.Trim().ToLower();
         
-        var userExists = await _repository.CheckExistenceAsync(u => u.Email == request.Email);
-        
+        var userExists = await _repository.CheckExistenceAsync(u => u.Email == req.Email);
         if (userExists)
             throw new ArgumentException("the email is already in use");
         
-        if (request.Username.Length < 3 || request.Username.Length > 10)
-            throw new ArgumentException("Username must be between 3 and 10 characters long");
-        if (!request.Email.Contains("@") || !request.Email.Contains("."))
-            throw new ArgumentException("Email must be a valid email address");
-        if (request.Password.Length < 8 || request.Password.Length > 22)
-            throw new ArgumentException("Password must be between 8 and 22 characters long");
-
+        _validator.Validate(req);
+        
         var newUser = new User
         {
-            Username = request.Username,
-            Email = request.Email,
-            Password = BC.HashPassword(request.Password, 6),
+            Username = req.Username,
+            Email = req.Email,
+            Password = BC.HashPassword(req.Password, 6),
         };
+        
         
         await _repository.AddAsync(newUser);
         await _cartService.AddAsync(new Cart { UserId = newUser.Id});
@@ -49,13 +48,17 @@ public class UserService : IUserService
     public async Task<UserDto> GetById(int id)
     {
         var user = await _repository.GetByIdAsync(id);
+        if (user == null)
+            throw new NotFoundException("No user found with the provided id");
         return _mapper.Map<UserDto>(user);
     }
     
     public async Task<UserDto> GetUserByEmail(string email)
     {
         email = email.Trim().ToLower();
-        var user = await _repository.GetFirstAsync(u => u.Email == email);
+        var user = await _repository.GetFirstOrDefaultAsync(u => u.Email == email);
+        if (user == null)
+            throw new NotFoundException("No user found with the provided email");
         return _mapper.Map<UserDto>(user);
     }
 
@@ -68,6 +71,8 @@ public class UserService : IUserService
     public async Task Delete(int id)
     {
         var user = await _repository.GetByIdAsync(id);
+        if (user == null)
+            throw new NotFoundException("No user found with the provided id");
         await _repository.DeleteAsync(user);
     }
 }
